requirements.txt
torch>=1.9.0
torchvision>=0.10.0
timm>=0.4.12
opencv-python>=4.5.0
numpy>=1.21.0
matplotlib>=3.4.0
scikit-learn>=1.0.0
Pillow>=8.3.0
tqdm>=4.62.0
kagglehub>=0.1.0